Removed HtsGetExporterController.
